requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
python-dateutil==2.9.0
twilio==9.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
crewai==0.80.0
crewai-tools==0.14.0
langchain-community==0.3.0
chromadb==0.5.0